[ui] Turn “prior events” link on asset details into a button, make it more obvious
Fixes #16601

Recently we’ve had several folks ask how to filter asset runs by partition because our partitions UI only shows the most recent run/materialization for each partition.

This PR converts the existing “(4 prior events)” link (that does what people are looking for) into a large “View all historical events (4)” button. It moves it to the far right of the partition details row so we don’t have to put it under either “run” or “latest materialization” and shows it all the time, disabling it rather than hiding it when there is nothing to display.

This PR also makes the parititon key on Asset > Events click through to the partitions page, so that folks trying to use that tab and filter by partition are more likely to find this Partitions page.